7.14 Bugfix

+++ src/controllers/videoController.js
@@ -26,15 +26,26 @@ export const getEdit = async (req, res) => {
     if (!video) {
         return res.status(404).render("404", {pageTitle: "Video not found."});
     }
+
+    const {user} = req.session;
+    if (String(video.owner) !== String(user._id)) {
+        return res.status(403).redirect("/");
+    }
+
     return res.render("edit", {pageTitle: `Edit: ${video.title}`, video});
 };
 
 export const postEdit = async (req, res) => {
     const {id} = req.params;
     const {title, description, hashtags} = req.body;
-    const existsVideo = await Video.exists({_id: id});
-    if (!existsVideo) {
+    const video = await Video.findById(id);
+    if (!video) {
         return res.render("404", {pageTitle: "Video not found."});
+    }
+
+    const {user} = req.session;
+    if (String(video.owner) !== String(user._id)) {
+        return res.status(403).redirect("/");
     }
 
     try {
@@ -82,6 +93,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const {id} = req.params;
+
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+
+    const {user} = req.session;
+    if (String(video.owner) !== String(user._id)) {
+        return res.status(403).redirect("/");
+    }
+
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
